initial code skeleton
