+++ src/App.jsx
@@ -1,26 +1,8 @@
-/*
- * Temporary problems array schema
- */
-const problems = [{
-    title: "201. Bitwise AND of Numbers Range",
-    difficulty: "Medium",
-    acceptance: "42%"
-},{
-    title: "201. Bitwise AND of Numbers Range",
-    difficulty: "Medium",
-    acceptance: "412%"
-},
-    {
-        title: "202. Happy Number",
-        difficulty: "Easy",
-        acceptance: "54.9%"
-    },
-    {
-        title: "203. Remove Linked List Elements",
-        difficulty: "Hard",
-        acceptance: "42%"
-    }];
-
+import problems from './problemList.js'
+import Login from "./components/login.jsx";
+import Signup from "./components/signup.jsx";
+import Problem from "./components/problem.jsx";
+import ProblemSet from "./components/problemSet.jsx";
 
 function App() {
 
@@ -31,29 +13,16 @@ function App() {
        /problems/:problem_slug - A single problem page
      */
 
+    console.log(problems)
     return (
     <div>
         Finish the assignment! Look at the comments in App.jsx as a starting point
+        <Login/>
+        <Signup/>
+        <ProblemSet/>
+        <Problem/>
     </div>
   )
 }
 
-// A demo component
-function ProblemStatement(props) {
-    const title = props.title;
-    const acceptance = props.acceptance;
-    const difficulty = props.difficulty;
-
-    return <tr>
-        <td>
-            {title}
-        </td>
-        <td>
-            {acceptance}
-        </td>
-        <td>
-            {difficulty}
-        </td>
-    </tr>
-}
 export default App
