+++ src/components/login.jsx
@@ -1,0 +1,9 @@
+function Login() {
+    return(
+        <div>
+            Login
+        </div>
+    )
+}
+
+export default Login
