+++ src/components/problem.jsx
@@ -1,0 +1,7 @@
+function Problem() {
+    return(
+        <div>Problem</div>
+    )
+}
+
+export default Problem
