+++ src/components/problemSet.jsx
@@ -1,0 +1,26 @@
+function ProblemSet(){
+    return (
+        <div>Problem Set</div>
+    )
+}
+
+// A demo component
+function ProblemStatement(props) {
+    const title = props.title;
+    const acceptance = props.acceptance;
+    const difficulty = props.difficulty;
+
+    return <tr>
+        <td>
+            {title}
+        </td>
+        <td>
+            {acceptance}
+        </td>
+        <td>
+            {difficulty}
+        </td>
+    </tr>
+}
+
+export default ProblemSet
