+++ src/components/signup.jsx
@@ -1,0 +1,7 @@
+function Signup() {
+    return(
+        <div>SignUp</div>
+    )
+}
+
+export default Signup
