+++ src/problemList.js
@@ -1,0 +1,24 @@
+/*
+ * Temporary problems array schema
+ */
+const problems = [{
+    title: "201. Bitwise AND of Numbers Range",
+    difficulty: "Medium",
+    acceptance: "42%"
+},{
+    title: "201. Bitwise AND of Numbers Range",
+    difficulty: "Medium",
+    acceptance: "412%"
+},
+    {
+        title: "202. Happy Number",
+        difficulty: "Easy",
+        acceptance: "54.9%"
+    },
+    {
+        title: "203. Remove Linked List Elements",
+        difficulty: "Hard",
+        acceptance: "42%"
+    }];
+
+export default problems
